Removing trailing slashes from node name in list command
Usually, when the user enters something like "/home/" a node
not found error should return but it is quite easy to fix, as
we do in this commit, just getting rid of the traling / symbols
at the end of the string.

Some unuseful vars in other commands were purged too.