Implementing commands `cd' and `pwd'
It is very useful to choose a node and exec commands against it
without having to type its whole name all the time. To do it, we
choosen to deal with it like we do with directories in an unix shell.

The `cd' command stores the given directory in a context attribute. At
this time, no command (not even `cd') is aware about this var. It will
be done in the next commits.

`cd' does not understand relative paths yet, if you are placed in the
/home directory, you'll not be able to cd to localhost. It will try to
find the absolute localhost node. I don't know how to deal with it
yet. It is because the node creation hierarchy is not a standard. All
tests I'm doing, I'm using the ejabberd's `hometree' implementation
but in the `flat' plugin things are different... we have to think =P