Adding first pieces of a GUI browser.
Now our browser has a pretty login form that collects
connection info and authentication credentials. Plus this,
we have started to work in the main window of our browser.
There is already a loading widget and a tiny API to
control its visibility, it will be useful to be used while
running async requests.