Adding the first pieces of the posts treeview and its feeding
This commit adds the necessary widgets to show the post list andhandles
the `row-activated' signal of the nodes treeview to load posts in the proper
widget.

There is nothing actually being appended to the post treview model
because I'm not sure that I should accept only atom content. Maybe a
configuration to make it possible to choose if the content is an atom or
just try to load it to an atom content, otherwise show it raw.