Proposing a new UI to PsBrowser
This new ui will count on three columns, one for bookmarks, another for
node lista and the laste for items of the selected node.

The idea is to bold selected entries in bookmarks and node list to show
to the user that it's selected.

I did not think about how to show the item content yet.