Adding an almost working node listing feature
To do it, I've added a method list_nodes that sends a pubsub query and
register the parse_list_nodes static method as the answer callback.

This anwer callback receives everything needed to show found results
in the node treeview. It currently has some problems with tree iters that
will be fixed soon.